#4.0 dev 修复分类排序值过大sql错误，分类别名为空报重复别名的bug
git-svn-id: http://emlog.googlecode.com/svn/trunk@1892 5f219224-4642-0410-9343-2d18db063b08

--HG--
branch : _bitbucket

+++ src/install.php
@@ -354,7 +354,7 @@ CREATE TABLE {$db_prefix}sort (
   sid tinyint(3) unsigned NOT NULL auto_increment,
   sortname varchar(255) NOT NULL default '',
   alias VARCHAR(200) NOT NULL DEFAULT '',
-  taxis tinyint(3) NOT NULL default '0',
+  taxis smallint(4) unsigned NOT NULL default '0',
   PRIMARY KEY  (sid)
 )".$add."
 DROP TABLE IF EXISTS {$db_prefix}trackback;
